test(tape): replace the JUCE-donor A/B with a tone regression
The core is Tape Machine 2 now and the donor's JUCE shell still carries
the v1 DSP, so nulling one against the other would assert that v2 sounds
like v1. What the master bus actually needs guarded is that v2's tone does
not drift after it lands, so the file renders a fixed program through
fixed settings and pins peak and RMS. Aggregate rather than sample-exact
because the suite runs on GCC, Clang, MSVC, x86_64 and arm64.

Latency is now a constant 56 samples at every oversampling factor, and
the factor itself is inert: the core's factorFromChoice is hardwired to
2x because the local anti-aliasing stages around each nonlinearity do the
work. Both are pinned rather than left as a surprise, since MasterBus
sizes its tape-crossfade dry delay from that latency report.

Dropping the donor half also drops its editor, its GUI translation units,
the createPluginFilter rename and the AnalogVUMeter collision workaround
that only existed because both donors' meters linked here.
